fix: add component and styles for product page
