+++ src/pages/ProductPage/ProductPage.tsx
@@ -1,18 +1,14 @@
-/* eslint-disable */
 import styled from '@emotion/styled';
-import { Box, Button, Rating, Typography } from '@mui/material';
+import { Box, Button } from '@mui/material';
 import React, { useEffect, useState } from 'react';
-import { useParams, Link } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import data from 'data';
-import { FC } from 'react';
-import ComponentWithChildren from 'types/ComponentWithChildren';
 import { ProductItem } from 'types/ProductItem';
 import { useAppDispatch, useAppSelector } from 'hooks/reduxHooks';
 import { addProductToCart, removeProductFromCart } from 'handlers/cartSlice';
-import currencyFormatter from 'helpers/currencyFormatter';
-import percentageFormatter from 'helpers/percentageFormatter';
 import BreadCrumbs from 'components/Breadcrumbs';
-import Gallery from 'components/Gallery';
+import ProductGallery from 'components/ProductGallery';
+import ProductDescription from 'components/ProductDescription';
 
 const PageContent = styled(Box)(() => ({
   margin: '10px auto',
@@ -36,17 +32,6 @@ const ProductText = styled(Box)(() => ({
   display: 'flex',
   flexDirection: 'column',
   alignItems: 'center',
-}));
-
-const ProductAbout = styled(Box)(() => ({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'flex-start',
-}));
-
-const Span = styled('span')(() => ({
-  fontWeight: '700',
-  fontSize: '1.1rem',
 }));
 
 const ButtonWrapper = styled(Box)(() => ({
@@ -86,41 +71,9 @@ const ProductPage = () => {
     <PageContent>
       <BreadCrumbs title={product.title} brand={product.brand} category={product.category} />
       <ProductContent>
-        <Gallery images={product.images} thumbnail={product.thumbnail} />
+        <ProductGallery images={product.images} thumbnail={product.thumbnail} />
         <ProductText>
-          <Typography variant="h5" fontWeight={600} textAlign="center" textTransform="capitalize">
-            {product.title}
-          </Typography>
-          <Typography
-            sx={{ maxWidth: '320px', textAlign: 'center', margin: '10px auto', fontSize: '1.1rem' }}
-          >
-            {product.description}
-          </Typography>
-          <ProductAbout>
-            <Typography>
-              <Span>Discount Percentage: </Span>
-              {percentageFormatter.format(product.discountPercentage / 100)}
-            </Typography>
-            <Typography sx={{ display: 'flex', alignItems: 'center' }}>
-              <Span>Rating: </Span>
-              <Rating name="read-only" max={5} value={product.rating} precision={0.5} readOnly />
-            </Typography>
-            <Typography>
-              <Span>Stock: </Span>
-              {product.stock}
-            </Typography>
-            <Typography>
-              <Span>Brand: </Span>
-              {product.brand}
-            </Typography>
-            <Typography>
-              <Span>Category: </Span>
-              {product.category}
-            </Typography>
-          </ProductAbout>
-          <Typography variant="h5" fontWeight={600} margin="20px 0 20px 0">
-            {currencyFormatter.format(product.price)}
-          </Typography>
+          <ProductDescription product={product} />
           <ButtonWrapper>
             {isProductInCart ? (
               <Button
